--- packages/next-loader-tag/src/errors.ts
@@ -1,2 +1,0 @@
-export const ERR_TAG_INVALID_INPUT = `"tag" 함수에 제공된 값이 유효한 태그 결과나 리졸버 함수와 일치하지 않습니다. 내부 오류입니다.`;
-export const ERR_EMPTY_TAG_ARRAY = `태그 배열은 비어 있을 수 없습니다. 최소 하나의 태그가 필요합니다.`;
